Extract brush editor export helpers
