+++ CMakeLists.txt
@@ -146,6 +146,7 @@ target_sources(
         src/fbx_loader.c
         src/game.c
         src/game_data_brush_artifacts.c
+        src/game_data_brush_editor_export.c
         src/game_data_brush.c
         src/game_data.c
         src/game_data_actions.c
